Revert "change data size from int64 to uint64 (#1497)" (#1540)
close pingcap/ticdc#1544

+++ downstreamadapter/dispatchermanager/helper.go
@@ -226,10 +226,7 @@ func (h *SchedulerDispatcherRequestHandler) Handle(eventDispatcherManager *Event
 	return false
 }
 
-func (h *SchedulerDispatcherRequestHandler) GetSize(event SchedulerDispatcherRequest) uint64 {
-	return 0
-}
-
+func (h *SchedulerDispatcherRequestHandler) GetSize(event SchedulerDispatcherRequest) int { return 0 }
 func (h *SchedulerDispatcherRequestHandler) IsPaused(event SchedulerDispatcherRequest) bool {
 	return false
 }
@@ -325,8 +322,8 @@ func (h *HeartBeatResponseHandler) Handle(eventDispatcherManager *EventDispatche
 	return false
 }
 
-func (h *HeartBeatResponseHandler) GetSize(event HeartBeatResponse) uint64 { return 0 }
-func (h *HeartBeatResponseHandler) IsPaused(event HeartBeatResponse) bool  { return false }
+func (h *HeartBeatResponseHandler) GetSize(event HeartBeatResponse) int   { return 0 }
+func (h *HeartBeatResponseHandler) IsPaused(event HeartBeatResponse) bool { return false }
 func (h *HeartBeatResponseHandler) GetArea(path common.GID, dest *EventDispatcherManager) int {
 	return 0
 }
@@ -378,8 +375,8 @@ func (h *CheckpointTsMessageHandler) Handle(eventDispatcherManager *EventDispatc
 	return false
 }
 
-func (h *CheckpointTsMessageHandler) GetSize(event CheckpointTsMessage) uint64 { return 0 }
-func (h *CheckpointTsMessageHandler) IsPaused(event CheckpointTsMessage) bool  { return false }
+func (h *CheckpointTsMessageHandler) GetSize(event CheckpointTsMessage) int   { return 0 }
+func (h *CheckpointTsMessageHandler) IsPaused(event CheckpointTsMessage) bool { return false }
 func (h *CheckpointTsMessageHandler) GetArea(path common.GID, dest *EventDispatcherManager) int {
 	return 0
 }
@@ -432,8 +429,8 @@ func (h *MergeDispatcherRequestHandler) Handle(eventDispatcherManager *EventDisp
 	return false
 }
 
-func (h *MergeDispatcherRequestHandler) GetSize(event MergeDispatcherRequest) uint64 { return 0 }
-func (h *MergeDispatcherRequestHandler) IsPaused(event MergeDispatcherRequest) bool  { return false }
+func (h *MergeDispatcherRequestHandler) GetSize(event MergeDispatcherRequest) int   { return 0 }
+func (h *MergeDispatcherRequestHandler) IsPaused(event MergeDispatcherRequest) bool { return false }
 func (h *MergeDispatcherRequestHandler) GetArea(path common.GID, dest *EventDispatcherManager) int {
 	return 0
 }
