+++ pkg/sink/mysql/helper.go
@@ -456,13 +456,13 @@ func queryMaxPreparedStmtCount(ctx context.Context, db *sql.DB) (int, error) {
 }
 
 // queryMaxAllowedPacket gets the value of max_allowed_packet
-func queryMaxAllowedPacket(ctx context.Context, db *sql.DB) (uint64, error) {
+func queryMaxAllowedPacket(ctx context.Context, db *sql.DB) (int64, error) {
 	row := db.QueryRowContext(ctx, "select @@global.max_allowed_packet;")
 	var maxAllowedPacket sql.NullInt64
 	if err := row.Scan(&maxAllowedPacket); err != nil {
 		return 0, cerror.WrapError(cerror.ErrMySQLQueryError, err)
 	}
-	return uint64(maxAllowedPacket.Int64), nil
+	return maxAllowedPacket.Int64, nil
 }
 
 func getDDLCreateTime(ctx context.Context, db *sql.DB) string {
